Add files via upload
The model still needs to be cleaned up and fixes need to be implemented. Currently the model is containing to many variables and dropping features seems to not work for me, also the VFR variable is not available once you scale the dataset. The residual analysis has not been performed due to the earlier issues, once resolved we can provide a better picture.